Add help screen, currently showing default shortcuts
Hitting '?' brings up a modal with default shortcuts, for quick 
reference.
Also does a little refactoring in the code where default shortcuts are 
declared.

Known issue: anyone that has run SH84 before will have shortcuts loaded 
from config file and the new '?' command will be ignored. Addressing 
this is beyond the scope of this change.

Known issue: custom shortcuts are not displayed in help, only the 
default which may not correspond to the actual user configuration.